+++ NEWS.md
@@ -1,3 +1,10 @@
+# insight 0.9.6
+
+## Changes to functions
+
+* `get_data()` now works for models from `afex_aov()`.
+* `get_parameters()` returns a more informative message for `BFBayesFactor` objects when not the first model is indexed.
+
 # insight 0.9.5
 
 ## Breaking changes